+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Implemented enhancements
 
+- no ticket - added second qualification form for invest email campaign
+
 ### Fixed bugs
 - GAA -13 - Added empty alt tag (tbh reverted)
 - GAA-27 - directory-components version bump
@@ -16,7 +18,7 @@
 - no ticket - directory-components version bump
 - GAA-26 - case study color contrast
 - GAA-24 - Added missing H1
-- GAA-32 - Fixed alt text on various template 
+- GAA-32 - Fixed alt text on various template
 - no ticket - alt tag added to invest template
 - TT-2236 Spanish Translation on great international
 - chore/python-gevent adding gevent asyncio for performance